% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_format.R
\name{ggformat}
\alias{ggformat}
\title{""}
\usage{
ggformat(plot, y_title, x_title, title = "", print = TRUE, size_text = 15)
}
\arguments{
\item{plot}{saved ggplot, variable name, must provide}

\item{y_title}{the title of y axis, must provide}

\item{x_title}{the title of x axis, must provide}

\item{title}{the title of the plot; default= no title}

\item{print}{logical, default = TRUE}

\item{size_text}{font size for text, default = 15}
}
\value{
formatted ggplot. Formatting themes include: axis, text size, label sizes, no gridlines, panel border around the plot
}
\description{
""
}
\examples{
plotToformat<-ggplot()+geom_point(mapping = aes(x = seq(0,10,0.1), y = sin(seq(0,10,0.1))))
ggformat(plotToformat, y = "Sine Wave", x = "Time", size_text = 15)
}
